Update target framework version .NET 9.0
