--- src/JsonEasyNavigation.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JsonEasyNavigation", "JsonEasyNavigation\JsonEasyNavigation.csproj", "{EFB42F15-6C46-4C1A-94D0-177516E8FDA7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JsonEasyNavigation.Tests", "JsonEasyNavigation.Tests\JsonEasyNavigation.Tests.csproj", "{5F8113CF-1160-4D4B-8E9A-20D36CD79F04}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EFB42F15-6C46-4C1A-94D0-177516E8FDA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFB42F15-6C46-4C1A-94D0-177516E8FDA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EFB42F15-6C46-4C1A-94D0-177516E8FDA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EFB42F15-6C46-4C1A-94D0-177516E8FDA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5F8113CF-1160-4D4B-8E9A-20D36CD79F04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F8113CF-1160-4D4B-8E9A-20D36CD79F04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F8113CF-1160-4D4B-8E9A-20D36CD79F04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F8113CF-1160-4D4B-8E9A-20D36CD79F04}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
