--- src/JsonEasyNavigation/JsonEasyNavigation.csproj
@@ -1,42 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFrameworks>net5.0;netstandard2.0</TargetFrameworks>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-        <Version>1.3.0</Version>
-        <PackageIcon>logo.png</PackageIcon>
-        <Title>JsonEasyNavigation</Title>
-        <Authors>Sharkadi Andrey</Authors>
-        <Description>This library provides a wrapper class around JsonElement (located in System.Text.Json) which allows to navigate through JSON DOM (domain object model) hierarchy using indexer-style syntax (as in collections and dictionaries) for properties and array alike. It also contains useful methods to get values without throwing exceptions. </Description>
-        <Copyright>(c) Sharkadi Andrey</Copyright>
-        <PackageProjectUrl>https://github.com/sharkadi-a/JsonEasyNavigation</PackageProjectUrl>
-        <PackageTags>text.json json navigation indexer dictionary list collection DOM</PackageTags>
-    </PropertyGroup>
-    <PropertyGroup>
-        <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include="..\..\media\logo.png" Pack="true" PackagePath="\" />
-    </ItemGroup>
-    
-    <ItemGroup>
-        <None Include="..\..\LICENSE.txt" Pack="true" PackagePath="" />
-    </ItemGroup>    
-    
-    <ItemGroup>
-        <None Include="..\..\README.md" Pack="true" PackagePath="" />
-    </ItemGroup>
-    
-    <ItemGroup>
-      <PackageReference Include="System.Text.Json" Version="5.0.2" Condition="'$(TargetFramework)' == 'netstandard2.0'" />
-    </ItemGroup>
-    
-    <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-      <DocumentationFile>bin\Debug\JsonEasyNavigation.xml</DocumentationFile>
-    </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-      <DocumentationFile>bin\Release\JsonEasyNavigation.xml</DocumentationFile>
-    </PropertyGroup>
-</Project>
