+++ src/README_BODY.md
@@ -1,60 +1,4 @@
-﻿# JsonEasyNavigation
-This library provides a wrapper class around JsonElement (located in System.Text.Json) which allows to navigate through JSON DOM (domain object model) hierarchy using indexer-style syntax (as in collections and dictionaries) for properties and array alike. It also contains useful methods to get values without throwing exceptions.
-
-
----
-
-
-<table>
-<tr>
-<th></th>
-<th>Description</th>
-</tr>
-<tr>
-<td>UPDATED:</td>
-<td>9/17/2024</td>
-</tr>
-<tr>
-<td>FRAMEWORK:</td>
-<td>netstandard2.0, netstandard2.1, net9.0</td>
-</tr>
-<tr>
-<td>LANGUAGE:</td>
-<td>[C#] preview</td>
-</tr>
-<tr>
-<td>OUTPUT TYPE:</td>
-<td>Library [API]</td>
-</tr>
-<tr>
-<td>SUPPORTS:</td>
-<td>[Visual Studio]</td>
-</tr>
-<tr>
-<td>GFX SUBSYS:</td>
-<td>[None]</td>
-</tr>
-<tr>
-<td>TAGS:</td>
-<td>[text.json json navigation indexer dictionary list collection DOM]</td>
-</tr>
-<tr>
-<td>STATUS:</td>
-<td><a href="https://github.com/Latency/JsonEasyNavigation/actions/workflows/dotnet.yml"><img src="https://github.com/Latency/JsonEasyNavigation/actions/workflows/dotnet.yml/badge.svg"></a></td>
-</tr>
-<tr>
-<td>LICENSE:</td>
-<td><a href="https://github.com/Latency/JsonEasyNavigation/blob/master/MIT-LICENSE.txt"><img src="https://img.shields.io/github/license/Latency/JsonEasyNavigation?style=plastic&logo=GitHub&logoColor=black&label=License&color=yellowgreen"></a></td>
-</tr>
-<tr>
-<td>VERSION:</td>
-<td><a href="https://github.com/Latency/JsonEasyNavigation/releases"><img src="https://img.shields.io/github/v/release/Latency/JsonEasyNavigation?include_prereleases&style=plastic&logo=GitHub&logoColor=black&label=Version&color=blue"></a></td>
-</tr>
-<!-- VERSION: 1.3.2 -->
-</table>
-
-
-![JsonEasyNavigation](media/logo.png "JsonEasyNavigation")
+![JsonEasyNavigation](src/Resources/project.png "JsonEasyNavigation")
 
 <hr>
 
@@ -163,12 +107,14 @@ This library can be installed using GitHub found [here](https://github.com/users
 
 <h2><a name="license">License</a></h2>
 
-This software is distributed under the Apache License 2.0. See [LICENSE].txt.
+The source code for the site is licensed under the MIT license, which you can find in
+the [MIT-LICENSE].txt file.
 
-All graphical assets are licensed under the [Creative Commons Attribution 3.0 Unported License](https://creativecommons.org/licenses/by/3.0/).
+All graphical assets are licensed under the
+[Creative Commons Attribution 3.0 Unported License](https://creativecommons.org/licenses/by/3.0/).
 
 [//]: # (These are reference links used in the body of this note and get stripped out when the markdown processor does its job.)
 
-   [Creative Commons Attribution 3.0 Unported License]: <https://creativecommons.org/licenses/by/3.0/>
-   [LICENSE]: <https://www.apache.org/licenses/LICENSE-2.0>
+   [GNU LESSER GENERAL PUBLIC LICENSE]: <http://www.gnu.org/licenses/lgpl-3.0.en.html>
    [MSDN article]: <https://msdn.microsoft.com/en-us/library/c5b8a8f9(v=vs.100).aspx>
+   [MIT-License]: <http://choosealicense.com/licenses/mit/>
